Initial commit for Botui QT5 and Tello Support
This commit is the initial commit incorporating Tello support
into Botui. Tello support is a work-in-progress.  With these
patches to Botui along with the Tello support patches for
libwallaby, one should be able to (1) start a Tello, (2) command
that Tello start transmitting video, (3) then go the video page
and see video output from the Tello.

QT5 Support

An artifact of adding Tello support is that Botui now uses QT5.
This seems to perform quite well.

Issues:

(1) Need to integrate the code that finds and commands the Tello
to start sending video into Botui.

(2) Several graphics such as sensor output, etc. do not appear.
Fixing these should be no problem for a QT expert.

(3) The Qt4->Qt5 conversion was done by hand w/o the Qt tools, so
there may be a couple of data structures that the Author did not
get quite right.

(4) This is not an all-inclusive list of issues, but we now have
a working example of Botui using Qt5.

Signed-off-by: Eugene Myers <cedarhouse@comcast.net>