Fixing Factory Page (#91)
The factory page was setup wrong, I thought I could just use the "factory widget" that was lost in production but it did not use the StandardWidget model. The UI would crash when opening the page, now it doesn't. I will have to make the buttons functional in a future release.